requirements.txt
# Core libraries
requests==2.31.0
PyYAML==6.0  # For reading config.yaml

# Testing framework
pytest==7.4.0
pytest-html==3.2.0  # For generating HTML reports

# Additional utilities
requests-mock==1.11.0  # For mocking API responses
pytest-xdist==3.3.1  # Run tests in parallel
